Added isLegalMove (easy stuff working) Added a hold first value and hold last value for each player class. (will be used for some rules)
Noticed some bugs, but we will fix these once we run the test cases (if not before)

We need to clean up some code.

I have not updated the project so I have not merged, your changes are not in this version so you may need to pull from the stack.